shipapp/requirements.txt
Django==2.0
djangorestframework==3.10.3
mysql-connector==2.2.9
mysqlclient==1.4.4
numpy==1.17.3
pandas==0.25.3
pkg-resources==0.0.0
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
six==1.12.0
